♻️ R6288F task: decompose context reindex projection
