+++ .agentplane/tasks/202605290610-R6288F/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 R6288F task: Context reindex projection decomposition [202605290610-R6288F]
