+++ .agentplane/tasks/202605290610-R6288F/pr/review.md
@@ -1,0 +1,38 @@
+# PR Review
+
+Created: 2026-05-29T06:10:23.294Z
+
+## Task
+
+- Task: `202605290610-R6288F`
+- Title: Context reindex projection decomposition
+- Status: DOING
+- Branch: `task/202605290610-R6288F/context-reindex-projection-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290610-R6288F/README.md`
+
+## Verification
+
+- State: ok
+- Note: Context reindex projection row helpers extracted into packages/agentplane/src/context/reindex-projection.ts; cmdContextReindex and readContextProjection semantics are preserved. Checks passed: bun test packages/agentplane/src/commands/context/release-readiness.test.ts (21 pass); bun run typecheck; bun run arch:check; bun run knip:check; bun run lint:core; bun run format:changed; bun run hotspots:check (runtime hotspots 4 -> 3).
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T06:10:23.294Z
+- Branch: task/202605290610-R6288F/context-reindex-projection-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../agentplane/src/context/reindex-projection.ts   | 242 ++++++++++++++++++++
+ packages/agentplane/src/context/reindex.ts         | 251 +--------------------
+ 2 files changed, 248 insertions(+), 245 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
